+++ src/backend/executor/execUtils.c
@@ -2076,7 +2076,6 @@ void mppExecutorFinishup(QueryDesc *queryDesc)
 		if (qeError)
 		{
 			estate->dispatcherState = NULL;
-			cdbdisp_destroyDispatcherState(ds);
 			ReThrowError(qeError);
 		}
