+++ src/test/regress/output/dispatch.source
@@ -30,6 +30,39 @@ select gp_wait_until_triggered_fault('fts_probe', 1, 1);
  Success:
 (1 row)
 
+-- Test cdbdisp_getDispatchResults() fail in CdbDispatchCopyStart()
+create table dispatch_tbl (a int) distributed by (a);
+insert into dispatch_tbl values (2),(1),(5);
+-- fault on seg1 to block insert command into the dispatch_tbl table
+select gp_inject_fault('DoCopyToFail', 'error', '', '', '',
+   1, 1, 0, dbid) from gp_segment_configuration
+   where content = 1 and role = 'p';
+ gp_inject_fault 
+-----------------
+ Success:
+(1 row)
+
+copy dispatch_tbl to stdout;
+ERROR:  fault triggered, fault name:'DoCopyToFail' fault type:'error'  (seg1 192.168.235.128:7003 pid=68980)
+--select gp_wait_until_triggered_fault('DoCopyToFail', 1, dbid)
+--  from gp_segment_configuration where content = 1 and role = 'p';
+select gp_inject_fault('DoCopyToFail', 'reset', dbid)
+   from gp_segment_configuration
+   where content = 1 and role = 'p';
+ gp_inject_fault 
+-----------------
+ Success:
+(1 row)
+
+select * from dispatch_tbl;
+ a 
+---
+ 1
+ 5
+ 2
+(3 rows)
+
+drop table dispatch_tbl;
 -- Test quoting of GUC values and database names when they're sent to segments
 set client_min_messages='warning';
 DROP DATABASE IF EXISTS "funny""db'with\\quotes";
